save_images.py
import base64
import os
from PIL import Image
import io

# Create temp directory if it doesn't exist
if not os.path.exists('temp'):
    os.makedirs('temp')

# Save the first image
with open('temp/image1.jpg', 'wb') as f:
    # Hard-coded base64 string representation of the first image would go here
    # This would normally come from the attachment content
    print("Please add the base64 content of the first image")
    
# Save the second image
with open('temp/image2.jpg', 'wb') as f:
    # Hard-coded base64 string representation of the second image would go here
    # This would normally come from the attachment content
    print("Please add the base64 content of the second image")

print("Images saved to temp directory") 