requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.2
Pillow>=8.0.0
scikit-learn>=0.24.0
pandas>=1.2.0
tqdm>=4.54.0
matplotlib>=3.3.0
deepface>=0.0.75
insightface>=0.6.0
opencv-python>=4.5.0
tensorflow>=2.4.0
tensorflow-gpu>=2.4.0
seaborn>=0.11.0
scipy>=1.5.0
onnxruntime-gpu>=1.10.0 